refs #699: consolidate resize movie-data route
